Refactoring
* Changed the closeDriver() method in the DriverSingleton class.
* Created the DropDownMenu class in which you need to implement the click() method.
* Refactored folders.
