+++ pom.xml
@@ -37,6 +37,12 @@
             <version>4.2.2</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>io.github.bonigarcia</groupId>
+            <artifactId>webdrivermanager</artifactId>
+            <version>4.2.2</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 
 </project>
