+++ src/main/java/content/AbstractContent.java
