+++ src/main/java/model/CategoryMenu.java
@@ -1,4 +1,4 @@
-package page;
+package model;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
