+++ src/main/java/model/DropDownMenu.java
@@ -1,0 +1,15 @@
+package model;
+
+import org.openqa.selenium.WebElement;
+
+import java.util.ArrayList;
+
+public class DropDownMenu {
+    private ArrayList<WebElement> menuItems;
+
+    public boolean click(){
+        //TODO ADD ARRAY INITIALIZATION
+        return false;
+    }
+
+}
