+++ src/main/java/model/Header.java
@@ -1,4 +1,4 @@
-package page;
+package model;
 
 import org.openqa.selenium.WebDriver;
