+++ src/main/java/model/TopNavBar.java
@@ -1,0 +1,13 @@
+package model;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class TopNavBar {
+    private WebElement contactUsLink;
+    private WebElement accountLink;
+    private ArrayList<WebElement> myAccountLinks;
+}
