+++ src/main/java/pages/BasePage.java
@@ -1,24 +1,26 @@
-package page;
+package pages;
 
 import driver.DriverSingleton;
+import model.CategoryMenu;
+import model.Header;
+import model.TopNavBar;
 import org.openqa.selenium.WebDriver;
 
-public abstract class AbstractBasePage {
-    protected WebDriver driver;
-    private String pageAddress;
+public abstract class BasePage {
     protected final int WAIT_TIMEOUT_SECONDS = 10;
+    private String pageAddress;
     public TopNavBar topNavBar;
     public Header header;
     public CategoryMenu categoryMenu;
 
 
-    protected AbstractBasePage(WebDriver driver) {
-        this.driver = driver;
+    protected BasePage() {
+
     }
 
 
     public void openPage() {
-        DriverSingleton.driver.get(this.pageAddress);
+        DriverSingleton.getDriver().get(this.pageAddress);
     }
 
 }
