--- src/test/java/page/TopNavBar.java
@@ -1,6 +1,0 @@
-package page;
-
-import org.openqa.selenium.WebDriver;
-
-public class TopNavBar {
-}
